chore: move packages/sendportal-core/tests to Root tests/
