+++ tests/TestCase.php
@@ -4,12 +4,26 @@ declare(strict_types=1);
 
 namespace Tests;
 
-use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
+use Illuminate\Support\Str;
+use Orchestra\Testbench\TestCase as BaseTestCase;
+use Sendportal\Base\SendportalBaseServiceProvider;
+use Sendportal\Base\Services\Messages\RelayMessage;
 
 abstract class TestCase extends BaseTestCase
 {
-    use CreatesApplication;
-    use TestSupportTrait;
+    use SendportalTestSupportTrait;
+
+    /**
+     * @param \Illuminate\Foundation\Application $app
+     * @return array
+     */
+    protected function getPackageProviders($app): array
+    {
+        return [
+            SendportalBaseServiceProvider::class,
+            SendportalTestServiceProvider::class,
+        ];
+    }
 
     /**
      * Setup the test environment.
@@ -20,7 +34,22 @@ abstract class TestCase extends BaseTestCase
 
         $this->withoutMix();
         $this->withExceptionHandling();
+        $this->mockRelayMessageService();
 
         $this->artisan('migrate')->run();
+    }
+
+    /**
+     * @return void
+     */
+    protected function mockRelayMessageService(): void
+    {
+        $service = $this->getMockBuilder(RelayMessage::class)
+            ->disableOriginalConstructor()
+            ->getMock();
+
+        $service->method('handle')->willReturn(Str::random());
+
+        app()->instance(RelayMessage::class, $service);
     }
 }
