+++ tests/Unit/Factories/MailAdapterFactoryTest.php
