+++ tests/Unit/Models/SubscriberTest.php
